requirements.txt
torch>=2.1.0
transformers>=4.36.0
pyarrow>=14.0.0
pyyaml>=6.0
numpy
